requirements.txt
fastapi
fuzzywuzzy
groq
httpx
pydantic
pymongo
python-dotenv
python-telegram-bot
python-telegram-bot
pytz
Requests
Telethon
uvicorn
python-multipart
python-Levenshtein